better error messages

+++ stats.go
@@ -178,7 +178,7 @@ UpConn:
 		if c.PID == uint(proc.Pid) {
 			for _, v := range rt.Config.Resources {
 				for _, r := range v {
-					if c.RemotePort == r.URL.Port {
+					if c.RemotePort == uint16(r.URL.Port) {
 						for _, ip := range ips[r.URL.Host] {
 							if ip.Equal(c.RemoteAddress) {
 								d++
@@ -210,7 +210,7 @@ func (rt *Runtime) LookUpResourceIps() map[string][]net.IP {
 	return ips
 }
 
-//log proc samples infinite loop
+// log proc samples infinite loop
 func (rt *Runtime) logRuntimeStats(proc *process.Process) {
 	go func() {
 		for {
