+++ upstream.go
@@ -4,14 +4,14 @@ import (
 	"strconv"
 )
 
-//URL describes host mapping
+// URL describes host mapping
 type URL struct {
 	Scheme string
 	Host   string
-	Port   uint16
+	Port   int
 }
 
-//String representation of our URL struct
+// String representation of our URL struct
 func (u URL) String() string {
 	return u.Scheme + "://" + u.Host + ":" + strconv.Itoa(int(u.Port))
 }
